:hammer: refactor: Client hpp/cpp 파일 구분 및 폴더 위치 수정
