+++ srcs/Client/Makefile
@@ -1,9 +1,11 @@
 CXX = c++
 
+GAME_ELEMENT_HEAD = includes/gameElements
+
+HEAD = -Iincludes -I${GAME_ELEMENT_HEAD}
 
 OPNE_GL = ../../mac_opengl
 CLIENT_SRCS_DIR = srcs
-HEAD = ${CLIENT_SRCS_DIR}/gameElements
 
 CLIENT_SRCS = $(wildcard ${CLIENT_SRCS_DIR}/*.cpp)
 
@@ -16,7 +18,7 @@ CLIENT_LFLAGS = -lglfw -lGL -lm -lXrandr -lXi -lX11 -lXxf86vm -lpthread -ldl -lX
 endif
 
 CLIENT_OBJS = $(CLIENT_SRCS:.cpp=.o)
-CLIENT_CFLAGS = -Wall -Wextra -Werror -std=c++11 -I${HEAD} -I ${OPNE_GL}/include/ -DGL_SILENCE_DEPRECATION
+CLIENT_CFLAGS = -Wall -Wextra -Werror -std=c++11 ${HEAD} -I ${OPNE_GL}/include/ -DGL_SILENCE_DEPRECATION
 CLIENT_NAME = pong_client
 
 DEBUG_FLAGS = -g -fsanitize=address
@@ -32,7 +34,7 @@ $(CLIENT_NAME): $(CLIENT_OBJS)
 	$(CXX) $(CLIENT_OBJS) $(LDFLAGS) $(CLIENT_LFLAGS) -o $(CLIENT_NAME)
 
 debug: $(CLIENT_SRCS)
-	$(CXX) $(CLIENT_CFLAGS) $(CLIENT_LFLAGS) $(DEBUG_FLAGS) $(CLIENT_SRCS) -I $(HEAD) -o $(CLIENT_NAME) 
+	$(CXX) $(CLIENT_CFLAGS) $(CLIENT_LFLAGS) $(DEBUG_FLAGS) $(CLIENT_SRCS) $(HEAD) -o $(CLIENT_NAME) 
 
 clean:
 	rm -f $(CLIENT_OBJS)
